+++ src/main/java/de/kosmos_lab/web/server/servlets/openapi/AsyncApiJSONServlet.java
@@ -4,6 +4,7 @@ import de.kosmos_lab.web.annotations.Operation;
 import de.kosmos_lab.web.annotations.responses.ApiResponse;
 import de.kosmos_lab.web.doc.openapi.ApiEndpoint;
 import de.kosmos_lab.web.doc.openapi.ResponseCode;
+import de.kosmos_lab.web.exceptions.ParameterNotFoundException;
 import de.kosmos_lab.web.server.WebServer;
 import de.kosmos_lab.web.server.servlets.BaseServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
@@ -13,7 +14,6 @@ import java.io.IOException;
 
 @ApiEndpoint(path = "/doc/asyncapi.json", userLevel = -1)
 public class AsyncApiJSONServlet extends AsyncApiYamlServlet {
-    public String cached = null;
 
     public AsyncApiJSONServlet(WebServer webServer) {
         super(webServer);
@@ -30,25 +30,28 @@ public class AsyncApiJSONServlet extends AsyncApiYamlServlet {
     )
 
     public void get(BaseServletRequest request, HttpServletResponse response) throws IOException {
-        if (cached == null) {
 
-
-            cached = parser.getJSON().toString(2);
-
-        }
 
         String host = null;
+        boolean export = request.getBoolean("export",false);
         try {
-            host = request.getRequest().getHeader("host");
-        } catch (Exception ex ) {
+            host = request.getParameter("host",false);
+        } catch (ParameterNotFoundException e) {
+            throw new RuntimeException(e);
+        }
+        if ( host == null ) {
+            try {
+                host = request.getRequest().getHeader("host");
+            } catch (Exception ex) {
 
+            }
         }
         if ( host != null ) {
-            sendJSON(request, response, server.replaceHostName(cached,host));
+            sendJSON(request, response, server.replaceHostName(server.getAsyncApiParser().getCachedJSON(),host));
 
         }
         else {
-            sendJSON(request, response, cached);
+            sendJSON(request, response, server.getAsyncApiParser().getCachedJSON());
         }