+++ src/main/java/de/kosmos_lab/web/server/servlets/openapi/OpenApiJSONServlet.java
@@ -4,6 +4,7 @@ import de.kosmos_lab.web.annotations.Operation;
 import de.kosmos_lab.web.annotations.responses.ApiResponse;
 import de.kosmos_lab.web.doc.openapi.ApiEndpoint;
 import de.kosmos_lab.web.doc.openapi.ResponseCode;
+import de.kosmos_lab.web.exceptions.ParameterNotFoundException;
 import de.kosmos_lab.web.server.WebServer;
 import de.kosmos_lab.web.server.servlets.BaseServlet;
 import de.kosmos_lab.web.server.servlets.BaseServletRequest;
@@ -11,16 +12,12 @@ import jakarta.servlet.http.HttpServletResponse;
 
 import java.io.IOException;
 
-import static de.kosmos_lab.web.server.servlets.openapi.OpenApiYamlServlet.parser;
 
 
 @ApiEndpoint(path = "/doc/openapi.json", userLevel = -1)
 public class OpenApiJSONServlet extends BaseServlet {
-    public String cached = null;
-
     public OpenApiJSONServlet(WebServer webServer) {
         super(webServer);
-
     }
 
     @Operation(
@@ -33,24 +30,21 @@ public class OpenApiJSONServlet extends BaseServlet {
     )
 
     public void get(BaseServletRequest request, HttpServletResponse response) throws IOException {
-        if (cached == null) {
 
+        String host = request.getParameter("host");;
+        if ( host == null ) {
+            try {
+                host = request.getRequest().getHeader("host");
+            } catch (Exception ex) {
 
-            cached = parser.getJSON().toString(2);
-
-        }
-        String host = null;
-        try {
-            host = request.getRequest().getHeader("host");
-        } catch (Exception ex ) {
-
+            }
         }
         if ( host != null ) {
-            sendJSON(request, response, server.replaceHostName(cached,host));
+            sendJSON(request, response, server.replaceHostName(server.getOpenApiParser().getCachedJSON(),host));
 
         }
         else {
-            sendJSON(request, response, cached);
+            sendJSON(request, response, server.getOpenApiParser().getCachedJSON());
         }