+++ src/main/java/de/kosmos_lab/web/server/servlets/openapi/OpenApiYamlServlet.java
@@ -4,6 +4,7 @@ import de.kosmos_lab.web.annotations.Operation;
 import de.kosmos_lab.web.annotations.responses.ApiResponse;
 import de.kosmos_lab.web.doc.openapi.ApiEndpoint;
 import de.kosmos_lab.web.doc.openapi.ResponseCode;
+import de.kosmos_lab.web.exceptions.ParameterNotFoundException;
 import de.kosmos_lab.web.server.OpenApiParser;
 import de.kosmos_lab.web.server.WebServer;
 import de.kosmos_lab.web.server.servlets.BaseServlet;
@@ -15,14 +16,11 @@ import java.io.IOException;
 
 @ApiEndpoint(path = "/doc/openapi.yaml", userLevel = -1)
 public class OpenApiYamlServlet extends BaseServlet {
-    static OpenApiParser parser = null;
-    public String cached = null;
+
 
     public OpenApiYamlServlet(WebServer webServer) {
         super(webServer);
-        if (parser == null) {
-            parser = new OpenApiParser(webServer);
-        }
+
 
     }
 
@@ -37,24 +35,22 @@ public class OpenApiYamlServlet extends BaseServlet {
     @Override
     public void get(BaseServletRequest request, HttpServletResponse response)
             throws IOException {
-        if (cached == null) {
 
+        String host = request.getParameter("host");;
 
-            cached = parser.getYAML();
+        if ( host == null ) {
+            try {
+                host = request.getRequest().getHeader("host");
+            } catch (Exception ex) {
 
-        }
-        String host = null;
-        try {
-            host = request.getRequest().getHeader("host");
-        } catch (Exception ex ) {
-
+            }
         }
         if ( host != null ) {
-            sendText(request, response, server.replaceHostName(cached,host));
+            sendTextAs(request, response, server.replaceHostName(server.getOpenApiParser().getCachedYAML(),host),"text/x-yaml");
 
         }
         else {
-            sendText(request, response, cached);
+            sendTextAs(request, response, server.getOpenApiParser().getCachedYAML(),"text/x-yaml");
         }
 
     }